requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
opencv-python-headless
numpy
